ai_modules/llm_agent-main/requirements.txt
fastapi
uvicorn
python-dotenv
pymysql
openai
requests
httpx
networkx
matplotlib
seaborn
tiktoken
langchain
langchain-openai
langgraph
pydantic-settings